// apps/frontend/frontend-app/generated-sources/ft-api-client/models/BenchmarkOutcome.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Fingertips API
 * An API to query public health indicator data.
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: ProfileFeedback@dhsc.gov.uk
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * benchmark outcome
 * @export
 */
export const BenchmarkOutcome = {
    NotCompared: 'NotCompared',
    Lowest: 'Lowest',
    Lower: 'Lower',
    Low: 'Low',
    Middle: 'Middle',
    High: 'High',
    Higher: 'Higher',
    Highest: 'Highest',
    Better: 'Better',
    Best: 'Best',
    Worse: 'Worse',
    Worst: 'Worst',
    Similar: 'Similar'
} as const;
export type BenchmarkOutcome = typeof BenchmarkOutcome[keyof typeof BenchmarkOutcome];


export function instanceOfBenchmarkOutcome(value: any): boolean {
    for (const key in BenchmarkOutcome) {
        if (Object.prototype.hasOwnProperty.call(BenchmarkOutcome, key)) {
            if (BenchmarkOutcome[key as keyof typeof BenchmarkOutcome] === value) {
                return true;
            }
        }
    }
    return false;
}

export function BenchmarkOutcomeFromJSON(json: any): BenchmarkOutcome {
    return BenchmarkOutcomeFromJSONTyped(json, false);
}

export function BenchmarkOutcomeFromJSONTyped(json: any, ignoreDiscriminator: boolean): BenchmarkOutcome {
    return json as BenchmarkOutcome;
}

export function BenchmarkOutcomeToJSON(value?: BenchmarkOutcome | null): any {
    return value as any;
}

export function BenchmarkOutcomeToJSONTyped(value: any, ignoreDiscriminator: boolean): BenchmarkOutcome {
    return value as BenchmarkOutcome;
}